Separated agent state out
